Se agrega función de exportación de registros en CVS
Se organiza código
