+++ src/app/pages/auditoria/components/ver-detalle-log-dialog/ver-detalle-log-dialog.component.css
@@ -149,3 +149,4 @@
   color: white;
   transform: scale(1.1);
 }
+
