+++ src/styles.css
@@ -24,9 +24,7 @@
     display: flex;
     justify-content: center; 
     align-items: center; 
-    margin-left: 10%;
-    margin-right: 10%;
-    width: auto;
+    width: 50%;
     padding: 5px;
 }
 
@@ -186,7 +184,7 @@ input, select, textarea{
 }
 
 .driver-popover.driverjs-theme .driver-popover-title {
-  color: #731514;
+  color: #731514 !important;
 }
 .driver-popover.driverjs-theme .driver-popover-description,
 .driver-popover.driverjs-theme .driver-popover-progress-text{
@@ -199,9 +197,9 @@ input, select, textarea{
 .driver-popover.driverjs-theme button {
   flex: 1;
   text-align: center;
-  background-color: #731514;
+  background-color: #731514 !important;
   color: #ffffff;
-  border: 2px solid #731514;
+  border: 2px solid #731514 !important;
   text-shadow: none;
   font-size: 14px;
   padding: 5px 8px;
@@ -209,7 +207,7 @@ input, select, textarea{
 }
 
 .driver-popover.driverjs-theme button:hover {
-  background-color: #3a0b0a;
+  background-color: #3a0b0a !important;
   color: #ffffff;
 }
 
@@ -230,17 +228,17 @@ input, select, textarea{
 }
 
 .driver-popover.driverjs-theme .driver-popover-arrow-side-left.driver-popover-arrow {
-  border-left-color: #731514;
+  border-left-color: #731514 !important;
 }
 
 .driver-popover.driverjs-theme .driver-popover-arrow-side-right.driver-popover-arrow {
-  border-right-color: #731514;
+  border-right-color: #731514 !important;
 }
 
 .driver-popover.driverjs-theme .driver-popover-arrow-side-top.driver-popover-arrow {
-  border-top-color: #731514;
+  border-top-color: #731514 !important;
 }
 
 .driver-popover.driverjs-theme .driver-popover-arrow-side-bottom.driver-popover-arrow {
-  border-bottom-color: #731514;
+  border-bottom-color: #731514 !important;
 }
